chore: Implemented City, Reactor and TerrainMap key methods

+++ src/main/java/org/example/Main.java
@@ -1,8 +1,8 @@
 package org.example;
 
 public class Main {
-    public static void main(String[] args) {
-        float[] mapSize = {400, 450};
+    public static void main(String[] args){
+        int[] mapSize = {400, 450};
         Simulation simulation = new Simulation(4, 2, mapSize);
         simulation.run();
     }
